+++ src/netprotocols/layer7/__init__.py
@@ -1,4 +1,4 @@
-from netprotocols.layer7.dhcp import DHCP
+from netprotocols.layer7.dhcp import DHCP, DHCPOption
 from netprotocols.layer7.dns import (
     DNS,
     DNSOverTCP,
@@ -11,6 +11,7 @@ from netprotocols.layer7.dns import (
 __all__ = [
     "DHCP",
     "DNS",
+    "DHCPOption",
     "DNSOverTCP",
     "DNSQuestion",
     "DNSResourceRecord",